escalation chains list fix, warning component has been added
